升级jdk和spring版本，以便于支持spring-ai

+++ src/main/java/org/bigmouth/gpt/controller/FavoritesController.java
@@ -57,7 +57,7 @@ public class FavoritesController {
         if (!Objects.equals(session.getUserId(), userId)) {
             return ResponseEntity.ok().build();
         }
-        int count = favoritesService.count(Wrappers.query(new Favorites().setUserId(userId).setSessionMessageId(messageId)));
+        long count = favoritesService.count(Wrappers.query(new Favorites().setUserId(userId).setSessionMessageId(messageId)));
         if (count > 0) {
             return ResponseEntity.ok().build();
         }
